[ALLUXIO-2691] Separate FileSystemMaster interface and impl
