+++ core/server/master/src/main/java/alluxio/web/WebInterfaceGeneralServlet.java
@@ -17,7 +17,9 @@ import alluxio.RuntimeConstants;
 import alluxio.StorageTierAssoc;
 import alluxio.master.AlluxioMasterService;
 import alluxio.master.block.BlockMaster;
+import alluxio.master.file.DefaultFileSystemMaster;
 import alluxio.master.file.FileSystemMaster;
+import alluxio.master.file.StartupConsistencyCheck;
 import alluxio.underfs.UnderFileSystem;
 import alluxio.util.FormatUtils;
 
@@ -168,7 +170,7 @@ public final class WebInterfaceGeneralServlet extends HttpServlet {
    */
   private void populateValues(HttpServletRequest request) throws IOException {
     BlockMaster blockMaster = mMaster.getMaster(BlockMaster.class);
-    FileSystemMaster fileSystemMaster = mMaster.getMaster(FileSystemMaster.class);
+    FileSystemMaster fileSystemMaster = mMaster.getMaster(DefaultFileSystemMaster.class);
 
     request.setAttribute("debug", Configuration.getBoolean(PropertyKey.DEBUG));
 
@@ -193,9 +195,9 @@ public final class WebInterfaceGeneralServlet extends HttpServlet {
     request.setAttribute("freeCapacity",
         FormatUtils.getSizeFromBytes(blockMaster.getCapacityBytes() - blockMaster.getUsedBytes()));
 
-    FileSystemMaster.StartupConsistencyCheck check = fileSystemMaster.getStartupConsistencyCheck();
+    StartupConsistencyCheck check = fileSystemMaster.getStartupConsistencyCheck();
     request.setAttribute("consistencyCheckStatus", check.getStatus());
-    if (check.getStatus() == FileSystemMaster.StartupConsistencyCheck.Status.COMPLETE) {
+    if (check.getStatus() == StartupConsistencyCheck.Status.COMPLETE) {
       request.setAttribute("inconsistentPaths", check.getInconsistentUris().size());
       request.setAttribute("inconsistentPathItems", check.getInconsistentUris());
     } else {
