+++ core/server/master/src/test/java/alluxio/master/file/DefaultFileSystemMasterTest.java
@@ -86,9 +86,9 @@ import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 
 /**
- * Unit tests for {@link FileSystemMaster}.
+ * Unit tests for {@link DefaultFileSystemMaster}.
  */
-public final class FileSystemMasterTest {
+public final class DefaultFileSystemMasterTest {
   private static final AlluxioURI NESTED_URI = new AlluxioURI("/nested/test");
   private static final AlluxioURI NESTED_FILE_URI = new AlluxioURI("/nested/test/file");
   private static final AlluxioURI NESTED_DIR_URI = new AlluxioURI("/nested/test/dir");
@@ -100,7 +100,7 @@ public final class FileSystemMasterTest {
   private CreateFileOptions mNestedFileOptions;
   private BlockMaster mBlockMaster;
   private ExecutorService mExecutorService;
-  private FileSystemMaster mFileSystemMaster;
+  private DefaultFileSystemMaster mFileSystemMaster;
   private long mWorkerId1;
   private long mWorkerId2;
 
@@ -155,7 +155,7 @@ public final class FileSystemMasterTest {
   }
 
   /**
-   * Tests the {@link FileSystemMaster#delete(AlluxioURI, DeleteOptions)} method.
+   * Tests the {@link DefaultFileSystemMaster#delete(AlluxioURI, DeleteOptions)} method.
    */
   @Test
   public void deleteFile() throws Exception {
@@ -206,7 +206,7 @@ public final class FileSystemMasterTest {
   }
 
   /**
-   * Tests the {@link FileSystemMaster#delete(AlluxioURI, DeleteOptions)} method with a
+   * Tests the {@link DefaultFileSystemMaster#delete(AlluxioURI, DeleteOptions)} method with a
    * non-empty directory.
    */
   @Test
@@ -227,7 +227,7 @@ public final class FileSystemMasterTest {
   }
 
   /**
-   * Tests the {@link FileSystemMaster#delete(AlluxioURI, DeleteOptions)} method for
+   * Tests the {@link DefaultFileSystemMaster#delete(AlluxioURI, DeleteOptions)} method for
    * a directory.
    */
   @Test
@@ -258,7 +258,7 @@ public final class FileSystemMasterTest {
   }
 
   /**
-   * Tests the {@link FileSystemMaster#getNewBlockIdForFile(AlluxioURI)} method.
+   * Tests the {@link DefaultFileSystemMaster#getNewBlockIdForFile(AlluxioURI)} method.
    */
   @Test
   public void getNewBlockIdForFile() throws Exception {
@@ -284,7 +284,7 @@ public final class FileSystemMasterTest {
   }
 
   /**
-   * Tests the {@link FileSystemMaster#getPersistenceState(long)} method.
+   * Tests the {@link DefaultFileSystemMaster#getPersistenceState(long)} method.
    */
   @Test
   public void getPersistenceState() throws Exception {
@@ -302,7 +302,7 @@ public final class FileSystemMasterTest {
   }
 
   /**
-   * Tests the {@link FileSystemMaster#getFileId(AlluxioURI)} method.
+   * Tests the {@link DefaultFileSystemMaster#getFileId(AlluxioURI)} method.
    */
   @Test
   public void getFileId() throws Exception {
@@ -328,7 +328,7 @@ public final class FileSystemMasterTest {
   }
 
   /**
-   * Tests the {@link FileSystemMaster#getFileInfo(AlluxioURI)} method.
+   * Tests the {@link DefaultFileSystemMaster#getFileInfo(AlluxioURI)} method.
    */
   @Test
   public void getFileInfo() throws Exception {
@@ -715,7 +715,7 @@ public final class FileSystemMasterTest {
 
   /**
    * Tests that an exception is in the
-   * {@link FileSystemMaster#createFile(AlluxioURI, CreateFileOptions)} with a TTL set in the
+   * {@link DefaultFileSystemMaster#createFile(AlluxioURI, CreateFileOptions)} with a TTL set in the
    * {@link CreateFileOptions} after the TTL check was done once.
    */
   @Test
@@ -752,7 +752,7 @@ public final class FileSystemMasterTest {
 
   /**
    * Tests that an exception is in the
-   * {@link FileSystemMaster#createDirectory(AlluxioURI, CreateDirectoryOptions)} with a TTL
+   * {@link DefaultFileSystemMaster#createDirectory(AlluxioURI, CreateDirectoryOptions)} with a TTL
    * set in the {@link CreateDirectoryOptions} after the TTL check was done once.
    */
   @Test
@@ -1035,8 +1035,8 @@ public final class FileSystemMasterTest {
   }
 
   /**
-   * Tests the {@link FileSystemMaster#setAttribute(AlluxioURI, SetAttributeOptions)} method and
-   * that an exception is thrown when trying to set a TTL for a directory.
+   * Tests the {@link DefaultFileSystemMaster#setAttribute(AlluxioURI, SetAttributeOptions)} method
+   * and that an exception is thrown when trying to set a TTL for a directory.
    */
   @Test
   public void setAttribute() throws Exception {
@@ -1102,7 +1102,7 @@ public final class FileSystemMasterTest {
   }
 
   /**
-   * Tests the {@link FileSystemMaster#rename(AlluxioURI, AlluxioURI, RenameOptions)} method.
+   * Tests the {@link DefaultFileSystemMaster#rename(AlluxioURI, AlluxioURI, RenameOptions)} method.
    */
   @Test
   public void rename() throws Exception {
@@ -1187,7 +1187,7 @@ public final class FileSystemMasterTest {
   }
 
   /**
-   * Tests {@link FileSystemMaster#free} on persisted file.
+   * Tests {@link DefaultFileSystemMaster#free} on persisted file.
    */
   @Test
   public void free() throws Exception {
@@ -1208,7 +1208,7 @@ public final class FileSystemMasterTest {
   }
 
   /**
-   * Tests {@link FileSystemMaster#free} on non-persisted file.
+   * Tests {@link DefaultFileSystemMaster#free} on non-persisted file.
    */
   @Test
   public void freeNonPersistedFile() throws Exception {
@@ -1221,7 +1221,7 @@ public final class FileSystemMasterTest {
   }
 
   /**
-   * Tests {@link FileSystemMaster#free} on pinned file when forced flag is false.
+   * Tests {@link DefaultFileSystemMaster#free} on pinned file when forced flag is false.
    */
   @Test
   public void freePinnedFileWithoutForce() throws Exception {
@@ -1236,7 +1236,7 @@ public final class FileSystemMasterTest {
   }
 
   /**
-   * Tests {@link FileSystemMaster#free} on pinned file when forced flag is true.
+   * Tests {@link DefaultFileSystemMaster#free} on pinned file when forced flag is true.
    */
   @Test
   public void freePinnedFileWithForce() throws Exception {
@@ -1258,7 +1258,7 @@ public final class FileSystemMasterTest {
   }
 
   /**
-   * Tests the {@link FileSystemMaster#free} method with a directory but recursive to false.
+   * Tests the {@link DefaultFileSystemMaster#free} method with a directory but recursive to false.
    */
   @Test
   public void freeDirNonRecursive() throws Exception {
@@ -1271,7 +1271,7 @@ public final class FileSystemMasterTest {
   }
 
   /**
-   * Tests the {@link FileSystemMaster#free} method with a directory.
+   * Tests the {@link DefaultFileSystemMaster#free} method with a directory.
    */
   @Test
   public void freeDir() throws Exception {
@@ -1292,7 +1292,8 @@ public final class FileSystemMasterTest {
   }
 
   /**
-   * Tests the {@link FileSystemMaster#free} method with a directory with a file non-persisted.
+   * Tests the {@link DefaultFileSystemMaster#free} method with a directory with a file
+   * non-persisted.
    */
   @Test
   public void freeDirWithNonPersistedFile() throws Exception {
@@ -1306,7 +1307,7 @@ public final class FileSystemMasterTest {
   }
 
   /**
-   * Tests the {@link FileSystemMaster#free} method with a directory with a file pinned when
+   * Tests the {@link DefaultFileSystemMaster#free} method with a directory with a file pinned when
    * forced flag is false.
    */
   @Test
@@ -1323,7 +1324,7 @@ public final class FileSystemMasterTest {
   }
 
   /**
-   * Tests the {@link FileSystemMaster#free} method with a directory with a file pinned when
+   * Tests the {@link DefaultFileSystemMaster#free} method with a directory with a file pinned when
    * forced flag is true.
    */
   @Test
@@ -1344,7 +1345,7 @@ public final class FileSystemMasterTest {
   }
 
   /**
-   * Tests the {@link FileSystemMaster#mount(AlluxioURI, AlluxioURI, MountOptions)} method.
+   * Tests the {@link DefaultFileSystemMaster#mount(AlluxioURI, AlluxioURI, MountOptions)} method.
    */
   @Test
   public void mount() throws Exception {
@@ -1448,7 +1449,7 @@ public final class FileSystemMasterTest {
   }
 
   /**
-   * Tests the {@link FileSystemMaster#stop()} method.
+   * Tests the {@link DefaultFileSystemMaster#stop()} method.
    */
   @Test
   public void stop() throws Exception {
@@ -1458,7 +1459,7 @@ public final class FileSystemMasterTest {
   }
 
   /**
-   * Tests the {@link FileSystemMaster#workerHeartbeat(long, List)} method.
+   * Tests the {@link DefaultFileSystemMaster#workerHeartbeat(long, List)} method.
    */
   @Test
   public void workerHeartbeat() throws Exception {
@@ -1569,9 +1570,9 @@ public final class FileSystemMasterTest {
     MasterRegistry registry = new MasterRegistry();
     JournalFactory factory = new MutableJournal.Factory(new URI(mJournalFolder));
     mBlockMaster = new BlockMaster(registry, factory);
-    mExecutorService =
-        Executors.newFixedThreadPool(2, ThreadFactoryUtils.build("FileSystemMasterTest-%d", true));
-    mFileSystemMaster = new FileSystemMaster(registry, factory,
+    mExecutorService = Executors
+        .newFixedThreadPool(2, ThreadFactoryUtils.build("DefaultFileSystemMasterTest-%d", true));
+    mFileSystemMaster = new DefaultFileSystemMaster(registry, factory,
         ExecutorServiceFactories.constantExecutorServiceFactory(mExecutorService));
 
     mBlockMaster.start(true);
