+++ core/server/master/src/test/java/alluxio/master/lineage/LineageMasterTest.java
@@ -20,6 +20,7 @@ import alluxio.job.CommandLineJob;
 import alluxio.job.Job;
 import alluxio.job.JobConf;
 import alluxio.master.MasterRegistry;
+import alluxio.master.file.DefaultFileSystemMaster;
 import alluxio.master.file.FileSystemMaster;
 import alluxio.master.file.options.CompleteFileOptions;
 import alluxio.master.journal.JournalFactory;
@@ -54,7 +55,7 @@ import java.util.concurrent.ThreadFactory;
  * Unit tests for {@link LineageMaster}.
  */
 @RunWith(PowerMockRunner.class)
-@PrepareForTest({FileSystemMaster.class})
+@PrepareForTest({DefaultFileSystemMaster.class})
 public final class LineageMasterTest {
   private ExecutorService mExecutorService;
   private LineageMaster mLineageMaster;
@@ -73,8 +74,8 @@ public final class LineageMasterTest {
     MasterRegistry registry = new MasterRegistry();
     JournalFactory factory =
         new MutableJournal.Factory(new URI(mTestFolder.newFolder().getAbsolutePath()));
-    mFileSystemMaster = Mockito.mock(FileSystemMaster.class);
-    registry.add(FileSystemMaster.class, mFileSystemMaster);
+    mFileSystemMaster = Mockito.mock(DefaultFileSystemMaster.class);
+    registry.add(DefaultFileSystemMaster.class, mFileSystemMaster);
     ThreadFactory threadPool = ThreadFactoryUtils.build("LineageMasterTest-%d", true);
     mExecutorService = Executors.newFixedThreadPool(2, threadPool);
     mLineageMaster = new LineageMaster(registry, factory,
