+++ core/server/master/src/test/java/alluxio/master/lineage/recompute/RecomputePlannerTest.java
@@ -14,6 +14,7 @@ package alluxio.master.lineage.recompute;
 import alluxio.job.CommandLineJob;
 import alluxio.job.Job;
 import alluxio.job.JobConf;
+import alluxio.master.file.DefaultFileSystemMaster;
 import alluxio.master.file.FileSystemMaster;
 import alluxio.master.file.meta.FileSystemMasterView;
 import alluxio.master.file.meta.PersistenceState;
@@ -35,7 +36,7 @@ import java.util.ArrayList;
  * Unit tests for {@link RecomputePlanner}.
  */
 @RunWith(PowerMockRunner.class)
-@PrepareForTest({FileSystemMaster.class})
+@PrepareForTest({DefaultFileSystemMaster.class})
 public final class RecomputePlannerTest {
   private RecomputePlanner mPlanner;
   private LineageStore mLineageStore;
@@ -49,7 +50,7 @@ public final class RecomputePlannerTest {
   public void before() {
     mLineageStore = new LineageStore(new LineageIdGenerator());
     mJob = new CommandLineJob("test", new JobConf("output"));
-    mFileSystemMaster = Mockito.mock(FileSystemMaster.class);
+    mFileSystemMaster = Mockito.mock(DefaultFileSystemMaster.class);
     Mockito.when(mFileSystemMaster.getFileSystemMasterView())
         .thenReturn(new FileSystemMasterView(mFileSystemMaster));
     mPlanner = new RecomputePlanner(mLineageStore, mFileSystemMaster);
